Extract backup to seperate package
